add stdlib=c++11 multithreading support, introduce uint256.h, linting

+++ PrimeFactors.h
@@ -1,7 +1,6 @@
 #ifndef BRUTE_PRIMEFACTORS_H
 #define BRUTE_PRIMEFACTORS_H
 
-#include <assert.h>
 #include <cstdlib>
 #include <iostream>
 #include <limits.h>
@@ -14,15 +13,16 @@ namespace brute {
     {
         public:
             /* Constructors, Destructors  */
-            PrimeFactors(unsigned int num);
+            PrimeFactors( unsigned int num );
             ~PrimeFactors();
 
             /* Functions */
-            unsigned int* get_factors() const
+            unsigned int*get_factors() const
             {
                 unsigned int factors[3] = { get_n(), get_p(), get_q() };
                 return factors;
             }
+
             inline unsigned int  get_n() const { return n; }
             inline unsigned int  get_p() const { return q; }
             inline unsigned int  get_q() const { return q; }
@@ -33,12 +33,12 @@ namespace brute {
 
         private:
             /* Variables  */
-            unsigned int n;
-            unsigned int p;
-            unsigned int q;
+            unsigned long long int n;
+            unsigned long long int p;
+            unsigned long long int q;
 
             /* Functions */
-            inline unsigned int* brute_force(unsigned int &num);
+            inline unsigned int* brute_force( unsigned int &num );
     };
 }
 #endif
